Add support for recording xsavec x86 instruction
This patch add support of the xsavec instruction. The logic is highly
similar to the logic of xsave. The "init optimization" is not handled in
this patch.

The values of X86_XSAVEC_* were obtained by testing for AVX and MPX, by
reading Intel x86 manuals, and looking at
arch/x86/include/asm/fpu/types.h in Linux source.

gdb/ChangeLog:
2018-10-05  Pierre Marsais <pierre.marsais@lse.epita.fr>

	* common/x86-xstate.h (X86_XSAVEC_AVX_SIZE,
	X86_XSAVEC_BNDREGS_SIZE, X86_XSAVEC_BNDCFG_SIZE,
	X86_XSAVEC_K_SIZE, X86_XSAVEC_ZMM_H_SIZE,  X86_XSAVEC_ZMM_SIZE,
	X86_XSAVEC_PKRU_SIZE, X86_XSAVEC_SIZE(XCR0)): New macros.
	* i386-tdep.c: (i386_process_record): Handle xsavec instruction.

gdb/testsuite/ChangeLog:
2018-10-05  Pierre Marsais <pierre.marsais@lse.epita.fr>

	* gdb.reverse/i386-xsave-reverse.c (xsavec_test): New function.
	(main): Use xsavec_test.
	* gdb.reverse/i386-xsave-reverse.exp: Test for xsavec
	instruction.